Pipeline and services refactor 5: Latex Generation and Supabase
